+++ RELEASE_NOTES.md
@@ -1,7 +1,15 @@
-### New in 0.4
+### New in 0.5
+
+### New in 0.4.353 (released 2015-05-05)
 
 * Breaking: `ValueObject` now uses public properties instead of both
   private and public fields
+* Breaking: Aggregate IDs are no longer `string` but objects implementing
+  `IIdentity`
+* Breaking: MSSQL transient exceptions are now retried
+* Breaking: All methods on `IMsSqlConnection` has an extra `Label` argument
+* New: `ITransientFaultHandler` added along with default retry strategies
+  for optimistic concurrency and MSSQL transient exceptions
 * New: Release notes added to NuGet packages
 * New: Better logging and more descriptive exceptions
 * Fixed: Unchecked missing in `ValueObject` when claculating hash